Move ResizeArrowFnType to element/types.ts
